Add IDL sugar for applying the required trait
Marking a structure member as required is a big deal, and it should be
something that's more visually apparent than the current `@required`
trait that can blend in with the rest of the traits applied to a
member.

This commit introduces sugar for applying the required trait to
structure members by suffixing a member with "!". This is semantically
equivalent to applying the trait using `@required` before the member or
using an `apply` statement with `@required`.

By applying the required trait using `!`, the vertical space needed to
define most structures should be reduced, making structures easier to
read.